fix: fix code according to code rewiew
