+++ .gitignore
@@ -106,3 +106,6 @@ dist
 # OS generated files #
 .DS_Store
 .DS_Store?
+
+# CSS build files
+src/sass/style.css
