+++ online-store/src/sass/abstract/_constants.scss
@@ -10,6 +10,8 @@ $color-grey: #cccccc;
 $color-dark-grey: #666;
 $color-semitransparent: hsla(0, 0%, 100%, 0.3);
 $color-overlay: rgba(80, 80, 80, 0.4);
+$color-reset: rgba(19, 20, 21, 0.3);
+$color-slider: rgba(25, 118, 210, 0.38);
 
 $content-width: 1600px;
 $pc-small-width: 1280px;
