+++ online-store/src/sass/components/_filters.scss
@@ -1,5 +1,5 @@
 .store-content__filter-group {
-  background-color: rgba(19, 20, 21, 0.3);
+  background-color: $color-reset;
   border: 1px solid $color-white;
   border-radius: 10px;
   backdrop-filter: blur(3px);
