+++ online-store/src/sass/components/_search.scss
@@ -1,6 +1,6 @@
 .store-content__search-window {
   backdrop-filter: blur(3px);
-  background-color: rgba(19, 20, 21, 0.3);
+  background-color: $color-reset;
   border: 1px solid $color-cyan;
   border-radius: 20px;
   padding-left: 20px;
