+++ online-store/src/sass/components/_slider.scss
@@ -16,7 +16,7 @@
 .year-slider {
   position: relative;
   border-radius: 5px;
-  background-color: rgba(25, 118, 210, 0.38);
+  background-color: $color-slider;
   height: 4px;
 }
 
@@ -57,7 +57,6 @@ input[type="range"]::-webkit-slider-thumb {
   pointer-events: auto;
   background-color: $color-blue;
   cursor: pointer;
-  z-index: 3;
   position: relative;
 }
